+++ sql/snetruth/sne_truth_variability_schema.sql
@@ -1,3 +1,3 @@
 CREATE TABLE sne_truth.sn_variability_truth
-                  (id TEXT, obsHistID INTEGER, MJD REAL, bandpass TEXT,
-                  delta_flux REAL);
+                  (id TEXT, obsHistID INTEGER, MJD DOUBLE PRECISION,
+		  bandpass TEXT, delta_flux REAL);
